+++ internal/fsx/freespace_unix.go
@@ -1,13 +1,13 @@
 //go:build !windows
 
-package syncengine
+package fsx
 
 import "golang.org/x/sys/unix"
 
-// availableDiskBytes returns the free space (available to unprivileged
+// FreeBytes returns the free space (available to unprivileged
 // users) on the filesystem holding dir. ok is false when it can't be
 // determined, in which case callers skip the free-space check.
-func diskFreeBytes(dir string) (bytes uint64, ok bool) {
+func FreeBytes(dir string) (bytes uint64, ok bool) {
 	var st unix.Statfs_t
 	if err := unix.Statfs(dir, &st); err != nil {
 		return 0, false